requirements.txt
gevent==1.0.2
greenlet==0.4.7
grequests==0.2.0
Pillow==2.8.2
requests==2.7.0
wheel==0.24.0
